refactor(LottoController): 클래스 변경 LottoGame -> LottoController
