+++ src/main/java/step3/LottoMain.java
@@ -1,10 +1,10 @@
 package step3;
 
-import step3.domain.LottoGame;
+import step3.controller.LottoController;
 
 public class LottoMain {
     public static void main(String[] args) {
-        LottoGame lottoGame = new LottoGame();
-        lottoGame.play();
+        LottoController lottoController = new LottoController();
+        lottoController.play();
     }
 }
